requirements.txt
Gunicorn
beautifulsoup4
blinker
datetime
Flask
numpy
python-dateutil
python-dotenv
requests
requests-cache
skyfield
splines
urllib3
apscheduler